sessions-app/requirements.txt
requests
boto3==1.14.28
botocore==1.17.28
PyJWT==1.7.1
schema==0.7.2
pytest
schematics==2.1.0
